refactor(review): types.go を型ごとのファイルに分割
- Comment + noEditingComment → comment.go
- Range → range.go
- Event + Label() → event.go
- InputMode → input_mode.go
- types_test.go → event_test.go にリネーム
- noEditingComment を ReviewState のある state.go へ移動
- Reset() の不要な reset() ラッパーを除去
- Range.Side/StartSide を string から gh.DiffSide に変更し不要なキャストを削除
- Event.Label() のドキュメントコメントを修正（表示用ラベルであることを明記）
