+++ internal/pr/review/input_mode.go
@@ -1,0 +1,10 @@
+package review
+
+// InputMode represents the active text-input state within the review drawer.
+type InputMode int
+
+const (
+	InputNone InputMode = iota
+	InputComment
+	InputSummary
+)
